[lldb][binding][python] migration to Python 3.0. fix Clion/Linux (KT-29623)

+++ llvmDebugInfoC/src/scripts/konan_lldb.py
@@ -128,11 +128,11 @@ class KonanHelperProvider(lldb.SBSyntheticValueProvider):
             evaluate("(int)Konan_DebugGetFieldType({}, {})".format(self._ptr, child)).GetValueAsUnsigned()
             for child in range(self._children_count)]
         self._children_type_addresses = [
-            long(evaluate("(void *)Konan_DebugGetFieldAddress({}, {})".format(self._ptr, index)).GetValue(), 0) for
+            int(evaluate("(void *)Konan_DebugGetFieldAddress({}, {})".format(self._ptr, index)).GetValue(), 0) for
             index in range(self._children_count)]
 
     def _read_string(self, expr, error):
-        return self._process.ReadCStringFromMemory(long(evaluate(expr).GetValue(), 0), 0x1000, error)
+        return self._process.ReadCStringFromMemory(int(evaluate(expr).GetValue(), 0), 0x1000, error)
 
     def _read_value(self, index):
         value_type = self._children_types[index]
@@ -179,7 +179,7 @@ class KonanStringSyntheticProvider(KonanHelperProvider):
         buff_addr = evaluate("(char *)Konan_DebugBuffer()").unsigned
 
         error = lldb.SBError()
-        s = self._process.ReadCStringFromMemory(long(buff_addr), int(buff_len), error)
+        s = self._process.ReadCStringFromMemory(int(buff_addr), int(buff_len), error)
         if not error.Success():
             raise DebuggerException()
         self._representation = s if error.Success() else fallback
